+++ src/deploy/functions/runtimes/discovery/v1alpha1.ts
@@ -44,6 +44,7 @@ export type WireEndpoint = build.Triggered &
   Partial<build.TaskQueueTriggered> &
   Partial<build.BlockingTriggered> &
   Partial<{ scheduleTrigger: WireScheduleTrigger }> & {
+    omit?: build.Field<boolean>;
     labels?: Record<string, string> | null;
     environmentVariables?: Record<string, string> | null;
     availableMemoryMb?: build.MemoryOption | build.Expression<number> | null;
@@ -63,7 +64,7 @@ export type WireEndpoint = build.Triggered &
     // We now use "serviceAccount" but maintain backwards compatability in the
     // wire format for the time being.
     serviceAccountEmail?: string | null;
-    region?: string[];
+    region?: build.ListField;
     entryPoint: string;
     platform?: build.FunctionsPlatform;
     secretEnvironmentVariables?: Array<ManifestSecretEnv> | null;
@@ -121,9 +122,10 @@ function parseRequiredAPIs(manifest: WireManifest): build.RequiredApi[] {
 function assertBuildEndpoint(ep: WireEndpoint, id: string): void {
   const prefix = `endpoints[${id}]`;
   assertKeyTypes(prefix, ep, {
-    region: "array",
+    region: "List",
     platform: (platform) => build.AllFunctionsPlatforms.includes(platform),
     entryPoint: "string",
+    omit: "Field<boolean>?",
     availableMemoryMb: (mem) => mem === null || isCEL(mem) || build.isValidMemoryOption(mem),
     maxInstances: "Field<number>?",
     minInstances: "Field<number>?",
@@ -395,6 +397,7 @@ function parseEndpointForBuild(
   copyIfPresent(
     parsed,
     ep,
+    "omit",
     "availableMemoryMb",
     "cpu",
     "maxInstances",