+++ src/test/database/removeRemote.spec.ts
@@ -7,7 +7,7 @@ import { RTDBRemoveRemote } from "../../database/removeRemote";
 describe("RemoveRemote", () => {
   const instance = "fake-db";
   const host = "https://firebaseio.com";
-  const remote = new RTDBRemoveRemote(instance, host);
+  const remote = new RTDBRemoveRemote(instance, host, /* disableTriggers= */ false);
   const serverUrl = utils.getDatabaseUrl(host, instance, "");
 
   afterEach(() => {
@@ -17,7 +17,7 @@ describe("RemoveRemote", () => {
   it("should return true when patch is small", () => {
     nock(serverUrl)
       .patch("/a/b.json")
-      .query({ print: "silent", writeSizeLimit: "tiny" })
+      .query({ print: "silent", writeSizeLimit: "tiny", disableTriggers: "false" })
       .reply(200, {});
     return expect(remote.deletePath("/a/b")).to.eventually.eql(true);
   });
@@ -25,7 +25,7 @@ describe("RemoveRemote", () => {
   it("should return false whem patch is large", () => {
     nock(serverUrl)
       .patch("/a/b.json")
-      .query({ print: "silent", writeSizeLimit: "tiny" })
+      .query({ print: "silent", writeSizeLimit: "tiny", disableTriggers: "false" })
       .reply(400, {
         error:
           "Data requested exceeds the maximum size that can be accessed with a single request.",
@@ -36,7 +36,7 @@ describe("RemoveRemote", () => {
   it("should return true when multi-path patch is small", () => {
     nock(serverUrl)
       .patch("/a/b.json")
-      .query({ print: "silent", writeSizeLimit: "tiny" })
+      .query({ print: "silent", writeSizeLimit: "tiny", disableTriggers: "false" })
       .reply(200, {});
     return expect(remote.deleteSubPath("/a/b", ["1", "2", "3"])).to.eventually.eql(true);
   });
@@ -44,11 +44,24 @@ describe("RemoveRemote", () => {
   it("should return false when multi-path patch is large", () => {
     nock(serverUrl)
       .patch("/a/b.json")
-      .query({ print: "silent", writeSizeLimit: "tiny" })
+      .query({ print: "silent", writeSizeLimit: "tiny", disableTriggers: "false" })
       .reply(400, {
         error:
           "Data requested exceeds the maximum size that can be accessed with a single request.",
       });
     return expect(remote.deleteSubPath("/a/b", ["1", "2", "3"])).to.eventually.eql(false);
+  });
+
+  it("should send disableTriggers param", () => {
+    const remoteWithDisableTriggers = new RTDBRemoveRemote(
+      instance,
+      host,
+      /* disableTriggers= */ true
+    );
+    nock(serverUrl)
+      .patch("/a/b.json")
+      .query({ print: "silent", writeSizeLimit: "tiny", disableTriggers: "true" })
+      .reply(200, {});
+    return expect(remoteWithDisableTriggers.deletePath("/a/b")).to.eventually.eql(true);
   });
 });