+++ src/test/emulators/storage/rules/config.spec.ts
@@ -31,6 +31,32 @@ describe("Storage Rules Config", () => {
     expect(result.content).to.contain("allow read, write: if true");
   });
 
+  it("should use default config for project IDs using demo- prefix if no rules file exists", () => {
+    const config = getOptions({
+      data: {},
+      path: resolvePath,
+    });
+    const result = getStorageRulesConfig("demo-projectid", config) as SourceFile;
+
+    expect(result.name).to.contain("templates/emulators/default_storage.rules");
+    expect(result.content).to.contain("allow read, write;");
+  });
+
+  it("should use provided config for project IDs using demo- prefix if the provided config exists", () => {
+    const rulesFile = "storage.rules";
+    const rulesContent = Buffer.from(StorageRulesFiles.readWriteIfTrue.content);
+    const path = persistence.appendBytes(rulesFile, rulesContent);
+
+    const config = getOptions({
+      data: { storage: { rules: path } },
+      path: resolvePath,
+    });
+    const result = getStorageRulesConfig("demo-projectid", config) as SourceFile;
+
+    expect(result.name).to.equal(path);
+    expect(result.content).to.contain("allow read, write: if true");
+  });
+
   it("should parse rules file for multiple targets", () => {
     const mainRulesContent = Buffer.from(StorageRulesFiles.readWriteIfTrue.content);
     const otherRulesContent = Buffer.from(StorageRulesFiles.readWriteIfAuth.content);