+++ templates/init/functions/typescript/index.ts
@@ -1,6 +1,6 @@
 import * as functions from "firebase-functions";
 
-// // Start writing Firebase Functions
+// // Start writing functions
 // // https://firebase.google.com/docs/functions/typescript
 //
 // export const helloWorld = functions.https.onRequest((request, response) => {